+++ src/core/iterator.py
