+++ src/scenarios/nguyen_convergence.py
@@ -2,7 +2,7 @@ import json
 import os
 import pickle
 
-from core.convergence import AlphaFlowIterator
+from core.iterator import AlphaFlowIterator
 from core.predictors.predictor_type import PredictorType
 from scenarios.nguyen_scenario import build_nguyen_network
 from scenarios.scenario_utils import get_demand_with_inflow_horizon
