+++ src/scenarios/sioux_falls_convergence.py
@@ -2,7 +2,7 @@ import json
 import os
 import pickle
 
-from core.convergence import AlphaFlowIterator
+from core.iterator import AlphaFlowIterator
 from core.predictors.predictor_type import PredictorType
 from importer.sioux_falls_importer import add_od_pairs, import_sioux_falls
 from scenarios.scenario_utils import get_demand_with_inflow_horizon
